Upgrade CloudFile.size field from int32 to uint64

+++ nucliadb_protos/rust/src/resources.rs
@@ -2,8 +2,8 @@
 pub struct CloudFile {
     #[prost(string, tag="1")]
     pub uri: ::prost::alloc::string::String,
-    #[prost(int32, tag="2")]
-    pub size: i32,
+    #[prost(uint64, tag="2")]
+    pub size: u64,
     #[prost(string, tag="3")]
     pub content_type: ::prost::alloc::string::String,
     #[prost(string, tag="4")]
